+++ lib/metrics_collector.rb
