+++ run_dark_blue.rb
@@ -10,7 +10,7 @@ require_relative "lib/bag_repository"
 require_relative "lib/bag_validator"
 require_relative "lib/data_transfer"
 require_relative "lib/dispatcher"
-require_relative "lib/jobs"
+require_relative "lib/metrics_collector"
 require_relative "lib/remote_client"
 require_relative "lib/repository_package_repository"
 require_relative "lib/status_event_repository"
